Create LICENSE.md
## Extended Description

### Citation Request
If you use this project in your work, the author requests the following citation format:
- Author: [Your Full Name]
- Year: [Year]
- Project Title: [Project Title]
- Repository: [Repository URL]

This citation helps acknowledge the work and effort that has gone into this project and supports the continuation of open project development.

### Ethical Use
This project is intended for use in [describe the intended use scope, such as "educational purposes", "scientific research", etc.]. Users are encouraged to consider the ethical implications of their applications, especially in scenarios that could directly affect individuals' lives or privacy.

### Fictitious Data
This project utilizes fictitious data. Users are reminded that the data contained in this project does not reflect real information and should not be used for decision-making based on reality.